requirements.txt
cryptography==42.0.8
Django==5.0.2
Pillow==10.4.0
python-dotenv==1.0.1
weasyprint==62.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
whitenoise==6.7.0
uvicorn==0.30.5